+++ src/controller/addonController.ts
@@ -1,92 +1,24 @@
 import * as extract from "extract-zip";
 import * as fs from "fs";
+import fetch from "node-fetch";
 import * as vscode from "vscode";
 
+import { makeAuthHeader } from "./credentialController";
+import { getRootFolderPath } from "./rootController";
 import constants from "../config/config";
+import { addToCache } from "../model/cache";
 import { addonInfoResponse, errorMessages } from "../types";
-import { addToCache } from "../utils/addonCache";
-import { makeAuthHeader } from "../utils/requestAuth";
-import { getRootFolderPath } from "../utils/reviewRootDir";
-import { getInput, getVersionChoice, promptOverwrite } from "../views/addonView";
+import {
+  getInput,
+  getVersionChoice,
+  promptOverwrite,
+} from "../views/addonView";
 import { promptProgress, showErrorMessage } from "../views/notificationView";
-
-
-async function fetchDownloadFile(fileId: string) {
-  const url = `${constants.downloadBaseURL}${fileId}`;
-  const headers = await makeAuthHeader();
-  const response = await fetch(url, { headers });
-  if (!response.ok) {
-    const errorMessages: errorMessages = {
-      window: {
-        404: `(Status ${response.status}): XPI download file not found`,
-        401: `(Status ${response.status}): Unauthorized request for XPI file`,
-        403: `(Status ${response.status}): Inadequate permissions for XPI file`,
-        other: `(Status ${response.status}): Could not fetch XPI file`,
-      },
-      thrown: {
-        404: "Download file not found",
-        401: "Unauthorized request",
-        403: "Forbidden request",
-        other: "Download request failed",
-      },
-    };
-
-    return await showErrorMessage(
-      errorMessages,
-      response.status,
-      fetchDownloadFile,
-      [fileId]
-    );
-  }
-  return response;
-}
-
-/**
- * Downloads and extracts an add-on to the root folder. If no add-on is specified, prompts the user.
- * @param urlGuid GUID of the add-on.
- * @param urlVersion The version to download.
- * @returns 
- */
-export async function downloadAndExtract(
-  urlGuid?: string,
-  urlVersion?: string
-) {
-  try {
-    const input = urlGuid || (await getInput());
-    const json: addonInfoResponse = await getAddonInfo(input);
-
-    const versionInfo = await getVersionChoice(input, urlVersion);
-    const addonFileID = versionInfo.fileID;
-    const version = versionInfo.version;
-    const guid = json.guid;
-    const addonID = json.id;
-
-    const workspaceFolder = await getRootFolderPath();
-    const compressedFilePath = `${workspaceFolder}/${guid}_${version}.xpi`;
-
-    await addToCache("reviewMeta", [guid], {
-      review_url: json.review_url,
-      file_id: addonFileID,
-      id: addonID,
-    });
-
-    await downloadAddon(addonFileID, compressedFilePath);
-
-    await extractAddon(
-      compressedFilePath,
-      `${workspaceFolder}/${guid}/${version}`
-    );
-    
-    return { workspaceFolder, guid, version };
-  } catch (error) {
-    console.error(error);
-  }
-}
 
 /**
  * Fetches version information for a given add-on.
  * @param input A string identifying a given add-on.
- * @param next 
+ * @param next tthe URL of the next batch of add-ons, if any
  * @returns The version information for an add-on.
  */
 export async function getAddonVersions(input: string, next?: string) {
@@ -128,6 +60,48 @@ export async function getAddonVersions(input: string, next?: string) {
 }
 
 /**
+ * Downloads and extracts an add-on to the root folder. If no add-on is specified, prompts the user.
+ * @param urlGuid GUID of the add-on.
+ * @param urlVersion The version to download.
+ * @returns
+ */
+export async function downloadAndExtract(
+  urlGuid?: string,
+  urlVersion?: string
+) {
+  try {
+    const input = urlGuid || (await getInput());
+    const json: addonInfoResponse = await getAddonInfo(input);
+
+    const versionInfo = await getVersionChoice(input, urlVersion);
+    const addonFileID = versionInfo.fileID;
+    const version = versionInfo.version;
+    const guid = json.guid;
+    const addonID = json.id;
+
+    const workspaceFolder = await getRootFolderPath();
+    const compressedFilePath = `${workspaceFolder}/${guid}_${version}.xpi`;
+
+    await addToCache("reviewMeta", [guid], {
+      review_url: json.review_url,
+      file_id: addonFileID,
+      id: addonID,
+    });
+
+    const writeStream = await downloadAddon(addonFileID, compressedFilePath);
+    await new Promise((resolve) => writeStream.on("finish", resolve));
+    await extractAddon(
+      compressedFilePath,
+      `${workspaceFolder}/${guid}/${version}`
+    );
+
+    return { workspaceFolder, guid, version };
+  } catch (error) {
+    console.error(error);
+  }
+}
+
+/**
  * Fetches the information about a given add-on.
  * @param input A string identifying a given addon.
  * @returns The addon information.
@@ -166,34 +140,76 @@ async function getAddonInfo(input: string): Promise<addonInfoResponse> {
 }
 
 /**
+ * Fetches and returns the response for the file fetch.
+ * @param fileId id of the xpi.
+ * @returns The response given by the API.
+ */
+async function fetchDownloadFile(fileId: string) {
+  const url = `${constants.downloadBaseURL}${fileId}`;
+  const headers = await makeAuthHeader();
+  const response = await fetch(url, { headers });
+  if (!response.ok) {
+    const errorMessages: errorMessages = {
+      window: {
+        404: `(Status ${response.status}): XPI download file not found`,
+        401: `(Status ${response.status}): Unauthorized request for XPI file`,
+        403: `(Status ${response.status}): Inadequate permissions for XPI file`,
+        other: `(Status ${response.status}): Could not fetch XPI file`,
+      },
+      thrown: {
+        404: "Download file not found",
+        401: "Unauthorized request",
+        403: "Forbidden request",
+        other: "Download request failed",
+      },
+    };
+
+    return await showErrorMessage(
+      errorMessages,
+      response.status,
+      fetchDownloadFile,
+      [fileId]
+    );
+  }
+  return response;
+}
+
+/**
  * Fetches and downloads the add-on to path.
  * @param fileId The fileId of the add-on's xpi
  * @param path The path to save the add-on to.
  */
-async function downloadAddon(fileId: string, path: string) {
+async function downloadAddon(fileID: string, filepath: string) {
+  const dest = fs.createWriteStream(filepath, { flags: "w" });
+  const handleError = async () => {
+    const errorMessages: errorMessages = {
+      window: {
+        other: `Could not download addon to ${filepath}`,
+      },
+      thrown: {
+        other: "Download failed",
+      },
+    };
+    await showErrorMessage(errorMessages, "other", downloadAddon, [
+      fileID,
+      filepath,
+    ]);
+    dest.close();
+  };
 
+  dest.on("error", handleError);
   promptProgress("Downloading Addon", async () => {
-      const response = await fetchDownloadFile(fileId);
-      const dest = fs.createWriteStream(path, { flags: "w" });
-      dest.write(await response.buffer());
-      dest.close();
+    try {
+      const response = await fetchDownloadFile(fileID);
+      const buffer = await response.buffer();
+      dest.write(buffer);
+      dest.end();
+    } catch (error) {
+      handleError();
+    }
+  });
 
-      if (!fs.existsSync(path)) {
-        const errorMessages: errorMessages = {
-          window: {
-              other: `Could not download addon to ${path}`,
-          },
-          thrown: {
-              other: "Download failed",
-          },
-        };
-
-        await showErrorMessage(errorMessages, "other", downloadAddon, [
-            fileId,
-            path,
-        ]);
-      }
-    });
+  return dest;
 }
 
 /**
@@ -207,10 +223,9 @@ async function extractAddon(
 ) {
   // if the directory exists, ask to overwrite.
   if (!(await dirExistsOrMake(addonVersionFolderPath))) {
-    try{
+    try {
       await promptOverwrite();
-    }
-    catch{
+    } catch {
       await fs.promises.unlink(compressedFilePath);
       return;
     }
@@ -219,7 +234,6 @@ async function extractAddon(
   await extract(compressedFilePath, {
     dir: addonVersionFolderPath,
   });
-
   await fs.promises.unlink(compressedFilePath);
 
   if (!fs.existsSync(addonVersionFolderPath)) {
