+++ src/controller/diffController.ts
@@ -1,7 +1,27 @@
 import { spawn } from "child_process";
 import * as vscode from "vscode";
 
-import { getDiffCommand } from "../utils/diffTool";
+import { promptDiffCommand } from "../views/diffView";
+
+
+export async function getDiffCommand() {
+  const config = vscode.workspace.getConfiguration("assay");
+  const diffCommand =
+    config.get<string>("diffTool") || (await setDiffCommand(config));
+  return diffCommand;
+}
+
+export async function setDiffCommand(config: vscode.WorkspaceConfiguration) {
+  try{
+    const input = promptDiffCommand();
+    await config.update("diffTool", input, true);
+    return input;
+  }
+  catch{
+    console.error("No diff command provided.");
+    return;
+  }
+}
 
 export async function openInDiffTool(uris: [vscode.Uri, vscode.Uri]) {
   const [left, right] = uris;
