+++ src/views/exportView.ts
@@ -5,8 +5,7 @@ import { QPOption } from "../types";
 export default async function getDeleteCommentsPreference() {
   const config = vscode.workspace.getConfiguration("assay");
   const savedPreference =
-    (config.get<string>("deleteCommentsOnExport") as QPOption) ||
-    QPOption.None;
+    (config.get<string>("deleteCommentsOnExport") as QPOption) || QPOption.None;
 
   if ([QPOption.Yes, QPOption.No].includes(savedPreference)) {
     return savedPreference === QPOption.Yes;
