+++ src/views/notificationView.ts
@@ -8,26 +8,28 @@ import { errorMessages } from "../types";
  * @param task The async task to perform.
  * @returns the result of the task.
  */
-export async function promptProgress(message: string, task: () => Promise<void>){
-    return await vscode.window.withProgress(
-        { title: "Assay", location: vscode.ProgressLocation.Notification },
-        async (progress) => {
-            progress.report({
-            message: message,
-            });
-            await task();
-        }
-    );
+export async function promptProgress(
+  message: string,
+  task: () => Promise<void>
+) {
+  return await vscode.window.withProgress(
+    { title: "Assay", location: vscode.ProgressLocation.Notification },
+    async (progress) => {
+      progress.report({
+        message: message,
+      });
+      await task();
+    }
+  );
 }
-
 
 /**
  * Displays an error message to the user, and the option to retry (if applicable).
- * @param errorMessages 
- * @param status 
- * @param tryAgainFunction 
- * @param tryAgainFunctionParams 
- * @returns 
+ * @param errorMessages
+ * @param status
+ * @param tryAgainFunction
+ * @param tryAgainFunctionParams
+ * @returns
  */
 export async function showErrorMessage(
   errorMessages: errorMessages,
@@ -51,7 +53,9 @@ export async function showErrorMessage(
     )
     .then((action) => {
       if (action?.title === tryAgainButton.title) {
-        return tryAgainFunction ? tryAgainFunction(...tryAgainFunctionParams) : Promise.resolve();
+        return tryAgainFunction
+          ? tryAgainFunction(...tryAgainFunctionParams)
+          : Promise.resolve();
       } else if (action?.title === fetchNewAddonButton.title) {
         // restart the process, but also throw an error to end the current process
         vscode.commands.executeCommand("assay.get");
